requirements-dev.txt
ipython==8.12.0
Pillow==9.3.0
crispy-tailwind==1.0.3
djangorestframework==3.15.2